Changed trigger suggestion to On Workstation Lock
Since the script still works on locking the device as opposed to unlock, it is a better suggestion. When using the lock trigger, the transition from the old background to the new one will appear seamless, since the background will have changed prior to login. This is a good improvement over the 1-3 second delay experienced using the unlock trigger.

This may also remove Spotlight tips immediately (if enabled) since it occurs before the logout completes, but will need to observe the next refresh to verify this.